ILU2-POO-TP2/src/controleur/ControlPrendreEtal.java
package controleur;

import villagegaulois.Village;
import personnages.Gaulois;

public class ControlPrendreEtal {
	private Village village;
	private ControlVerifierIdentite controlVerifierIdentite;

	public ControlPrendreEtal(ControlVerifierIdentite controlVerifierIdentite,
			Village village) {
		this.controlVerifierIdentite = controlVerifierIdentite;
		this.village = village;
	}

	public boolean resteEtals() {
		//TODO a completer, attention le retour ne dit pas etre false :-)
		
		return this.village.rechercherEtalVide();
	}

	public int prendreEtal(String nomVendeur, String produit, int nbProduit) {
		//TODO a completer
		
		Gaulois gaulois = this.village.trouverHabitant(nomVendeur);
		return this.village.installerVendeur(null, produit, nbProduit);
		//int numeroEtal = -1;
		//return numeroEtal;
	}

	public boolean verifierIdentite(String nomVendeur) {
		//TODO a completer, attention le retour ne dit pas etre false :-)
		return this.controlVerifierIdentite.verifierIdentite(nomVendeur);
	}
}
